Unittest/Integrationtest for services done.

+++ src/backend/OneTime.Core/Services/Implementations/TimeEntryService.cs
@@ -49,6 +49,11 @@ public class TimeEntryService : ITimeEntryService
 
     public async Task ReplaceTimeEntries(int timesheetId, List<TimeEntry> newEntries)
     {
+	    if (timesheetId <= 0)
+	    {
+		    throw new ArgumentException("TimesheetId must be greater than zero");
+	    }
+	    
         await _timeEntryRepository.DeleteEntriesByTimesheetId(timesheetId);
         await _timeEntryRepository.AddTimeEntries(newEntries);
         await _timeEntryRepository.SaveChangesAsync();
@@ -74,9 +79,9 @@ public class TimeEntryService : ITimeEntryService
 	public async Task<IEnumerable<TimeEntry>> GetTimeEntriesForTimesheet(int userId, int timesheetId)
 	{
 		if (userId <= 0) 
-			throw new ArgumentException("UserId must be greater than zero");
+			throw new ArgumentException("UserId must be greater than zero.");
 		if (timesheetId <= 0) 
-			throw new ArgumentException("TimeSheetId must be greater than zero");
+			throw new ArgumentException("TimeSheetId must be greater than zero.");
 
 		var entries = await _timeEntryRepository.GetWeeklyTimeEntriesByUser(userId, timesheetId);
